slice compress and test

+++ basic/comparator.h
@@ -69,7 +69,7 @@ class Comparator<std::string> {
 public:
   
   bool Less(const std::string& lhs, const std::string& rhs) {
-    for (int i = 0; i < std::min(lhs.size(), rhs.size()); i++) {
+    for (size_t i = 0; i < std::min(lhs.size(), rhs.size()); i++) {
       if (lhs[i] != rhs[i]) {
         if (lhs[i] < rhs[i]) return true;
         else return false;
